+++ component/site/views/icals/tmpl/export.php
@@ -420,7 +420,7 @@ if (!empty($this->icalEvents))
 					$stamptime = JevDate::strftime("%Y%m%dT%H%M%S", time());
 					$originalstart = JevDate::strftime("%Y%m%dT%H%M%S", $originalstart);
 				}
-				$html .= "DTSTAMP$tzid:" . $stamptime . "\r\n";
+				$html .= "DTSTAMP:" . $stamptime . "\r\n";
 				$html .= "DTSTART$tzid:" . $chstart . "\r\n";
 				$html .= "DTEND$tzid:" . $chend . "\r\n";
 				$html .= "RECURRENCE-ID$tzid:" . $originalstart . "\r\n";